test(ts): adjust persistent headings tests
